build: complete warning-free build at -Wall -Wextra -Wpedantic -Werror
Clear the remaining warnings across disp3D, all applications (mne_browse,
mne_analyze, mne_analyze_studio, mne_scan, mne_inspect, mne_align), examples
and testframes so the whole project builds warning-free at the strict level.

Key decisions per review:
* Kept -Wpedantic strict but silenced only -Wvariadic-macro-arguments-omitted:
  it is triggered by Qt's own QSKIP(statement, ...) macro at 900+ legitimate
  call sites, not by our code, and the project already targets C++20.
* Preserved information: constants and helpers that document a layout, protocol,
  step index, role enum or operation are marked [[maybe_unused]] rather than
  deleted; only meaningless scratch locals were removed.

Real defects found and fixed by the offensive:
* enterEvent(QEvent*) hid QWidget::enterEvent(QEnterEvent*) (Qt6 signature).
* InvPwlRapMusic::calculateInverse hid then (after a using-decl) ambiguously
  overloaded the base virtual — now a proper override with matching signature.
* fScale passed uninitialised as the initial scale to fitMatchedPoints in
  ex_coreg and test_coregistration.

Categories otherwise fixed: unused variables/parameters/functions/constants/
typedefs, dead set-but-unused counters, pessimizing moves, ctor member-init
order, sign-compare, unused lambda captures, and QDateTime Qt::UTC deprecation.

+++ src/applications/mne_analyze/libs/anShared/Model/eventmodel.cpp
@@ -200,7 +200,7 @@ QVariant EventModel::data(const QModelIndex &index,
 
     auto events = m_EventManager.getEventsInGroups(m_selectedEventGroups);
 
-    if(index.row() >= events->size())
+    if(index.row() >= static_cast<int>(events->size()))
         return QVariant();
 
     if (index.isValid()) {
@@ -507,7 +507,7 @@ MatrixXi EventModel::getEventMatrix()
     auto events = m_EventManager.getEventsInGroups(m_selectedEventGroups);
 
     matEventDataMatrix.resize(events->size(), 3);
-    for (int i = 0; i < events->size(); i++){
+    for (int i = 0; i < static_cast<int>(events->size()); i++){
         matEventDataMatrix(i,0) = events->at(i).sample;
         matEventDataMatrix(i,1) = 0;
         matEventDataMatrix(i,2) = 1;
